Move op_tracker and perf_plotter to the component folder
