+++ test/cppsuite/src/component/op_tracker.cpp
