+++ test/cppsuite/src/component/perf_plotter.cpp
