+++ test/cppsuite/src/component/perf_plotter.h
