+++ test/cppsuite/src/main/test.cpp
@@ -30,7 +30,7 @@
 
 #include "src/common/api_const.h"
 #include "src/common/logger.h"
-#include "src/common/perf_plotter.h"
+#include "src/component/perf_plotter.h"
 
 namespace test_harness {
 test::test(const test_args &args) : _args(args)
